requirements.txt
python-telegram-bot
motor
httpx
fastapi
uvicorn